// src/app/data/models/ao3-symbols.enum.ts
export enum ContentRating {
  None = "symbols/none.png",
  General = "symbols/rating/rating-general-audience.png",
  Teen = "symbols/rating/rating-teen.png",
  Mature = "symbols/rating/rating-mature.png",
  Explicit = "symbols/rating/rating-explicit.png",
}

export enum RPO {
  None = "symbols/none.png",
  FF = "symbols/RPO/category-femslash.png",
  FM = "symbols/RPO/category-het.png",
  Gen = "symbols/RPO/category-gen.png",
  MM = "symbols/RPO/category-slash.png",
  Multi = "symbols/RPO/category-multi.png",
  Other = "symbols/RPO/category-other.png",
}

export enum ContentWarning {
  None = "symbols/none.png",
  Unspecified = "symbols/warning/warning-choosenotto.png",
  Explicit = "symbols/warning/warning-yes.png",
  External = "symbols/warning/warning-external-work.png",
}

export enum Status {
  Unknown = "symbols/none.png",
  InProgress = "symbols/status/complete-no.png",
  Completed = "symbols/status/complete-yes.png",
}